[FIX] sign: tone on tone invisible button in the sign app
Steps to reproduce:

Install website and choose the kiddo theme
Install Sign
In Sign, click on the "Share" of "Non disclosure Agreement.pdf"
In another window (not log in), access the shared sign document. Make sure to be in "mobile" mode in this view (using the browser dev tools)
Choose a field to fill in the document
The space under the field value looks empty rather than having the "Next" button

Issue:
The next button is there but it's color is purple and the background is purple too -> barely visible

Solution:
We change the color of the next and final button to white so that the writing becomes visible.

opw-2871042

closes odoo/enterprise#30260

X-original-commit: ef8ddee2b06a2feb46e3fa0eed23729ef3f2eba5
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>