[FIX] web_enterprise: mobile: scroll in selection mode
On a mobile device:
- Go to Inventory > Product Variants
- Switch to the list view
- Long press on a record
-> We can't scroll to select records below the current screen

When in selection mode, we want to prevent every user interaction with
the widget fields. This means preventing clicks on them.

`preventDefault()` was then called on `touchstart` events.
This also means that other touch events are cancelled such as scoll
events.

By registering event listeners in the capture phase, we can intercept
and cancel events before they arrive to the unwanted widgets.

Here's an example list view with the record selector made visible, a priority
widget and a label.

.-------------------------------.
|.------------.                 |
|| 1 selected |                 |
|'------------'                 |
|.-----------------------------.|
|| ☐     ★     Screw           ||
|'-----------------------------'|
|.-----------------------------.|
|| ☑     ★     Bolt            ||
|'-----------------------------'|
'-------------------------------'

In selection mode, when clicking on a star widget, `_onClickCapture` is
called before the click handler of the `Priority widget`.

If the event target is not the record selector
`.o_list_record_selector`, we stop its propagation.

We then search the current row by traversing up the DOM tree.
Finally, we go down the DOM to trigger a click event on the record
selector.

We also cancel the mouseover and mouseout events, as some widgets also
uses that.

opw-2860688

closes odoo/enterprise#30308

X-original-commit: de9fa5dbcb6a4d52b871ec1946fcfd4ddeb8c7fa
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>