[FIX] stock_barcode: remove product_id @api.depends on stock_move_line
The field product_stock_quant_ids is used specifically for 1 view as a readonly field.
As so, it does not need to be recomputed each time the product quants are updated.

Removing the @api.depends result in a big performance gain for products tracked by SN with a number of quants and a number of sml > to 100 000+.
In the scenario above, validating a transfer with 1000 units of a product tracked by SN would result in 1000 sql queries equivalent to "SELECT id FROM stock_move_line WHERE product_id = XXXX"; the computation time is ~ 0.5 second when this query returns 100 000+ smls; multiply this by 1000, and you reach a few minutes to validate the transfer.
The query above is generated by a search on the _modified_trigger function.

OPW-2893131

closes odoo/enterprise#36316

X-original-commit: ee738ee20692187477b7b9bc0053e9fccd2ae3ca
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: David <dafr@odoo.com>