[FIX] account_intrastat: unknown country code inconsistency fix
Utilise different unknown country codes depending on the country for
which the report is being generated.
----
European geonomenclature provides a series of country codes. These
generally follow the ISO standard for the two-letter coutry codes for
existing countries. There are also other more sybolic codes, among
which are the uknown country codes.

https://ec.europa.eu/eurostat/en/web/products-manuals-and-guidelines/-/ks-gq-20-010
(see page 27 in the pdf)
As specified in the EU geonomenclature (linked above), the unknown
country code is to be:
QU - Countries and territories not specified
QV - Countries and territories not specified within the framework of
intra-EU trade

The issue is that, depending on the authority, the system accepts one of
these, or the other.

In Belgium, for instance, QU is the expected value and QV is rejected,
resulting in an error (see screenshot in PR
https://github.com/odoo/enterprise/pull/30072)

In the Netherlands QV is expected, and QU will be rejected, resulting in
an error (see screenshot PR
https://github.com/odoo/enterprise/pull/29227).

Thus the solution is to utilise a dictionary with the company's country
code as the key, and the unknown country code as the value

closes odoo/enterprise#30445

Ticket-id: 2922206
X-original-commit: 1d29a91d17027c69819411807604dfd96d81dec3
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>