[FIX] voip: PhoneField: do not crash on click
Before this commit, it crashed when the user clicked on the value
of a PhoneField, with voip installed. After this commit, it opens
the voip DialingPanel as expected.

This commit also duplicates and adapts the (small) test suite for
the voip PhoneField, s.t. the new implementation is tested as well.
Before, only the legacy implementation was tested.

closes odoo/enterprise#30064

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>