[IMP] Sign: Send all mail with one company
At the moment, when working in a multi-company environment,
the sign invitation email is sent with the current company of the
 user (good), but all others (refusal/completion email) are sent
with the main company of the "inviter" - which may not be the same.

All emails should use the same company:
the one the user was in when they sent the request.

closes odoo/enterprise#30077

Task-id: 2882486
Signed-off-by: Arnaud Joset <arj@odoo.com>