[FIX] stock_barcode: unfold grouped line button
This commit fixes a little visual bug.
On the Stock Barcode App, if a line is grouped (for tracked products),
a button is displayed to show or hide the sublines.
This button's icon changes depending of if sublines are displayed or not
but before this commit, it always has the same icon.

Part-of: odoo/enterprise#27547