[FIX] l10n_nl_reports: speedup general_ledger xaf report
An OFFSET was introduced in the get_xaf query to avoid
a MemoryError. This works well for small databases but
when working with big databases, the OFFSET becomes
a performance bottleneck. This is because
even though OFFSET skips rows, those skipped rows
must still be computed inside the server. So a large offset
can become quite ineffecient.

This can be seen by looking at the cost estimation of psql's planner.
With large offsets the cost skyrockets and the planner moves
from Nested Joins to Hash Joins, which are much less efficient
for the get_xaf query.

To solve this issue, the OFFSET is replaced by a ROW_NUMBER
filter. The ROW_NUMBER window function will add a number column
to the resulting set of a query. Then, by adding a condition
in the WHERE clause, we can paginate the results of the query.

This approach has the advantage of being independent of the
number of skipped rows.

opw-2703534

closes odoo/enterprise#31569

X-original-commit: 599c6badd89d32ddfc96bc79863d719f6e5777b7
Signed-off-by: Laurent Smet <las@odoo.com>