[FIX] account: fix bank statement import from BNP
Before this commit, importing an .OFX file from BNP failed because the account number format is not the same.

This is a problem on BNP's side but it is blocking for our clients and needs to be fixed.

This commit fixes this by checking that the account number corresponds to a specific part of the IBAN (for France only).

Task id 2860720

closes odoo/enterprise#30149

X-original-commit: 871edb2c2aaadf77dec7b2b39cad8759d2078324
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>