[FIX] l10n_cl_edi: for foreign customers, prevent the need to use
55555555-5 as the partner vat if the customer is foreigner, allowing also
to use its own original vat number.
In this case, we evaluate whether the partner is a foreign customer using the l10n_cl_taxpayer_type field, which must have the value '4'. This is because, if we evaluate if the client is foreign using only the country_id, we run the risk of preventing the possibility of creating a conventional ticket (boleta) if that client does not qualify (or does not need) an invoice (export invoice code 110).

closes odoo/enterprise#30863

X-original-commit: 996d79d5f2fea6d4240663ea0370f984bde4286b
Signed-off-by: Josse Colpaert <jco@odoo.com>