[FIX] l10n_mx_edi_40: Payment Report
The CFDI for the payment with version 4 now has the element "pago20"
instead of the "pago10" that the 3.3 had. So in the report, the info
was being retrieved from the "pago10" which doesn't exist with
version 4.0

Now if there is no "pago10" element it will search for the "pago20"
element.

Part-of: odoo/enterprise#31000