[IMP] sign: add reply_to in sign.request emails
Currently in sign, when emails for a sign request are sent, no reply_to param
is set on the mail.mail record. So if someone replies to the email, it will end
nowhere. This commit adds the email of the person who requested the signature,
so that if anyone replies to the email, the responsible person will receive it
and will be able to continue the conversation.

task-2702011

closes odoo/enterprise#30134

Signed-off-by: Arnaud Joset <arj@odoo.com>