[IMP] *: removes the task_id field from the planning app
* = project_forecast, project_timesheet_forecast,
project_timesheet_forecast_contract, project_timesheet_forecast_sale,
sale_project_forecast

In this commit we have edited all the models and views of the planning app
in order to remove the task_id field and its dependencies.
Along with that, we have removed/modified other fields/views in other apps
(project, timesheets) that are related to the planning app's task_id field.
This is done in order to simplify the planning app for the end users, by
reducing the granularity of their planning info.

task-2941828

closes odoo/enterprise#30404

Related: odoo/odoo#98035
Related: odoo/upgrade#3771
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>