[IMP] web: map: add group by many2many
The purpose of this commit is to enable the "group by" functionality of
many2many in map views.
Some other views also implement it (e.g. Gannt odoo/enterprise#20192,
List and Kanban #74985), it is therefore a logical step to also support group
by many2many in map views.

The many2many 'group by' was available for the map view but
was not handled.

Resulting in
- The 'grouped by' map view with no group label
- The grouped by was by set of ids in the x2many and thus
 not grouped by each id in the x2many.
- The pin in the map had IDs that do not exist in the list of records on the right

Example: Project->My Tasks -> Map View -> Group by assignee
- There is no label for each users
- there is a "group by" for the set ["Mitchell Admin", "Marc Demo"]
 instead to find the same record in the group record of Mitchell
and also in the group record of Marc
- There is a pin record with the number 52 but this number does
not exist anymore in the list on the right.

task_id: 2721035

closes odoo/enterprise#30272

Related: odoo/odoo#98009
Signed-off-by: Michaël Mattiello <mcm@odoo.com>