[FIX] voip_onsip: don't cast possible FieldCommand to boolean
The default value of areCredentialsSet is false, meaning that if the
compute returns the field command `clear()`, we would expect the value to
be reset to false. However, casting the `clear` command to a boolean
would result in true, which is the opposite of what is expected.

This commit pulls the call to super from the boolean cast so that
possible field commands as return value aren't wrongly converted to
true.

closes odoo/enterprise#30164

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>