[REF] account_invoice_extract: OWL extends legacy
As the account and account_accountant views gets converted to OWL, the
account_invoice_extract module has to extends the previous legacy views and
not the freshly converted ones.

Also, we need to disable the account_invoice_extract tests as the migration
makes them fail.
Those tests are supposed to be activated back when the migration of
account_invoice_extract is done.

Part-of: odoo/enterprise#30076