[IMP] {website_}appointment{_hr}: Replace work hours by anytime
Purpose:

Give users a better experience when using the appointment,
website_appointment and appointment_hr modules.

Specifications:

Changing the "work hours" appointment type to an "anytime"
appointment type making it easy for the user to share a link
which includes all of his available time (only during work hours
if appointment_hr installed).

Changing the website_appointment simplified add form by the
standard form. This brings the user a better customization
when creating an appointment type.

Adding a pivot view to the appointment to give the user an
overview.

Auto-publish and redirect on the appointment when it's created
from the website.

Other modifications are mainly wording and views improvements.
See attachment file in parent task for the complete list.

Task-2900764

UPG PR: odoo/upgrade#3705

closes odoo/enterprise#29756

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>