[FIX] account_sepa: allow SEPA credit transfers for Swiss companies
In commit [1], we removed the warning messages for Swiss companies
because they were not relevant. However, it was still possible to
generate european credit transfers if all conditions were met.

In commit [2], we refactored and improved the way the errors were
displayed with a dedicated wizard instead of a UserError message.
However, we forced the use of generic credit transfers for Swiss
companies at the same time.

This behavior is not correct as Switzerland is part of the Single Euro
Payments Area, it should therefore be possible to generate SEPA payments
for Swiss companies.

[1]:https://github.com/odoo/enterprise/commit/643f9d64d8677a1a7bf50ef9ff24867f2899c152
[2]:https://github.com/odoo/enterprise/commit/b5d0420cb27aaef225c77b5f2b07657f8b9866a8

closes odoo/enterprise#30657

X-original-commit: 32d21e4b314746c0b9c700c465a21ed17e1e7153
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>