[FIX] helpdesk_ticket: fix calculation of "SLA Failed" flag
Before this commit, when the reached date in SLA status is earlier
than (or equal to) the deadline than the ticket could have SLA
reached failed if the deadline is before the current date because
the condition which checks if the deadline is before the current
date is not only done when the reached date is not defined.

This commit fixes the issue by checking if the reached date is not
set before executing the condition to check if the deadline is
earlier than the current date.

Steps to reproduce the issue:
----------------------------

1. We have SLA policy with deadline (assume 5 min after creation
   of ticket).
2. Our ticket met conditions for SLA (conditions were met before
   5 minutes, so "SLA Failed" Flag is false).
3. After SLA deadline (let us say 15 minutes after ticket creation)
   the Kanban stage of ticket changes.

Current behavior:
----------------
The SLA is reached as failed in that ticket.

Expected behavior:
-----------------
The SLA should be successful reached since the reached date is
earlier than the deadline.

opw-2869202

closes odoo/enterprise#30584

X-original-commit: 521f8eceb5d5308123f6f0ef9478a44c7157496f
Signed-off-by: Xavier <xbo@odoo.com>