[FIX] l10n_be_intrastat: weight to 2 decimal places
The specifications for the Belgian intrastat xml state that the weight
should be given in kilograms, with a maximum of 2 decimal places. This
commit ensures that the weight values in the xml are represented to 2
decimal places.

Although the precision on the weight field is specified by Stock Weight
decimal.precision model, which would in most cases be 2, if the
precision is otherwise defined, it can result in an invalid declaration.

The solution is to update the weight values of the lines in the xml.

closes odoo/enterprise#30888

Ticket-id: 2862157
X-original-commit: 22f433777a60b6753ad34fb0082c3d2427e6237a
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>