[FIX] documents_account: synchronize partner all the time
A big issue with the "create bill" action from document is that the
partner is not synced unless everything is configured properly in the
settings as far as documents > Accounting goes and the journal/main
attachment is changed.
However there is a clear issue with partners not being synchronized on
the documents.

Assume the following flow:
 - Bill is uploaded to the documents app
 - Create Bill action is used to create the account.move
 - Information on the bill is added partner, date, etc

The document will now also follow any change on the partner.

TaskId-2751063

closes odoo/enterprise#30694

X-original-commit: 3f53e50989fbb9551565429aede7c0a1e4d6e5a4
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>