[IMP] spreadsheet_edition: remove some context menu items
Remove some context menu items to reduce cluttering.
Adapt the tests to use the menu items in the topbar instead.

There is now a "data" category in the topbar menu of o_spreadsheet,
append the pivot/list menu items to it instead of remplacing the category.

Odoo task 2936630

Part-of: odoo/enterprise#30367