[FIX] project_enterprise: increase the query counter
Before this commit, the query counter fails with the fix done in
odoo/odoo#97501, because this fix adds additional query if the data is
not in the cache during the installation of project_enterprise module.

This commit adds a query to avoid any issue in the performance because
of the fix done in community.

Related PR: odoo/odoo#97501

closes odoo/enterprise#30322

X-original-commit: d22fb024ed1003236de237c48befec10a283e493
Related: odoo/odoo#97523
Signed-off-by: Xavier <xbo@odoo.com>