[FIX] quality_mrp{,_workorder}: define the lot on QC
In some cases, the lot of a component is defined with the producing lot

To reproduce the issue:
(Need quality_mrp_workorder)
1. Create two storable and tracked-by-usn products P_finished and
P_compo
2. Update the quantity of P_compo: 1 x USN_COMPO
3. Create a bill of materials:
    - Product: P_finished
    - Components: 1 x P_compo
    - Operations: add a basic one OP
4. Add a step to OP:
    - Type: Register Consumed Materials
    - Product To Register: P_compo
5. Create and confirm a MO with 1 x P_finished
6. Edit the MO:
    - Lot/Serial Number: USN_FINISHED
7. Open the tablet view of OP

Error: The serial number for P_compo is USN_FINISHED instead of
USN_COMPO

The incorrect lot is stored in the field `lot_id` of a WO:
https://github.com/odoo/enterprise/blob/e543ab013dae6540075b2f0c2e319caa9e2a8afc/mrp_workorder/models/mrp_workorder.py#L54
As shown, the error comes from the lot defined on the QC. Step 6, when
setting the producing lot, it triggers `_compute_lot_line_id`. This
compute will define the lot of the QC with the producing lot of the MO.
However, the conditions are not restrictive enough: in case of a
consumed materials registering, we should not use the producing lot.

OPW-2948027

closes odoo/enterprise#30338

X-original-commit: 26a595d8cd864b40b7c00f13e63358d059a72720
Signed-off-by: Adrien Widart <awt@odoo.com>
Signed-off-by: Tiffany Chang <tic@odoo.com>