[IMP] web_studio: studio approval 🤝 next activities
Before this commit, it was impractical for a "manager" to know what
approval flows were waiting for them, leaving users to manually contact
their manager on a record asking for approval of the action. This made
the feature less useful in general.

This commit introduces the notion of a "responsible" on an approval rule
that will be notified that an approval is waiting for action via a next
activity that is "linked" to the approval (so granting or refusing the
approval will automatically remove the activity, and marking the
activity as done will automatically approve the action if it hasn't
been approved or rejected already).

To keep a clean link between approval rules, records and next activities
that pertain to them, a new model 'studio.approval.request' is
introduced which models such a link (indeed, since the rule has no entry
yet, there was no clean model on which to set this link and adding a
field on activities for that was not a good idea, as it would have too
wide an impact).

Task-2895596

Part-of: odoo/enterprise#30349