[FIX] helpdesk: compute `partner_ticket_count` correctly
Before this commit, the partner_ticket_count computation was providing
a wrong value as it was always reducing partner tickets count by 1
(unconditionally whether current ticket was present in that list or not).

With this commit, the number of tickets is only reduced by 1 if the current
ticket is part of them.

task-2927335

closes odoo/enterprise#33351

X-original-commit: 45d89dd412a8314843fd2008b0ea7df817b6e722
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Xavier <xbo@odoo.com>