[FIX] knowledge: let admins invite new users without write access
When the administrator does not have 'write' access on an article, the
administrator will not be able to invite new users.

To reproduce the error:

As "Marc Demo" (demo user):
1. Create an article
2. Invite "Mitchell Admin" on the article and set his permission to "read"

As "Mitchell Admin" (admin user):
1. Open the article you created with "Marc Demo"
2. Open the permission panel
3. Fill the modal form with a random member
4. Click on the "invite" button.

This commit will update the ir.rule of the `knowledge.article.member`
model to let the administrators invite new users without needing the
"write" permission on the article.

task-2893614

X-original-commit: abceff03269b2ae56c66af342cdd894f33840fcb
Part-of: odoo/enterprise#30430