[FIX] spreadsheet: avoid formula loading flickering
In a spreadsheet with lots of different data sources, we can often see
formulas flickering between "Loading..."/empty cell/#ERROR before finally
resolving to the final value.

Currently, `LoadingDataError` is raised only when the (pivot/list/etc.) model has not been
created yet. When it's been created but the data is not loaded yet, all pivot
formulas resolves to an empty string. Only when the data is loaded it finally
takes the real value.

With a basic ODOO.PIVOT formula, we quickly have:
1. model not created, fetching metadata: "Loading..."
2. model is loading: ""
3. the final value: 123

With relational ODOO.PIVOT.HEADER, we have:
1. model not created, fetching metadata: "Loading..."
2. model is loading: ""
3. name_get to fetch related record name: "Loading.."
4. the final value: "Azure Interior"

With a formula `=A1/A2` where both references are ODOO.PIVOT formulas:
1. model not created, fetching metadata: "Loading..."
2. model is loading, both cell values are an empty string: "#ERROR"
   => you can't divide strings)
3. the final value: 123

Task 2906465

Part-of: odoo/enterprise#30208