[FIX] account_followup: modifying the template doesn't modify history
How to reproduce the bug ?

- install base and account_accountant
- In the Accounting app, go to Customers > Follow-up Reports. Open one
report.
- If no report has been sent, send one.
- Click on the body of the message and edit it.
- Click on save and check the report that you have already sent

What is the bug ?

When you modify the body of the follow-up report and sav it, the
non-editable text area must be updated. When the javascript function
selects the text area, it will also fetch the text area of the reports
sent and visible in the chatter.

opw-2939986

closes odoo/enterprise#30353

X-original-commit: 4df6fffed5131fa0022b33c14e29a4895fbd3642
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>