[FIX] knowledge: correctly build toc
When a first section is started with smaller heading, immediately followed by a
bigger heading, the bigger heading is displayed under the first smaller one
inside toc, which should not be the case. For example, consider the following
HTML in an article:

<h3>Header 1<h3>
<h2>Header 2</h2>

The toc for above HTML is being built like:

Header 1
---- Header 2

This is wrong because h2 being bigger than h3, should also be at 0 depth
instead of 1.

This commit fixes the issue and builds the toc correctly for above mentioned
case.

taskID-2893657

closes odoo/enterprise#30310

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>