[IMP] social_*: improve post preview performance
Technical note: the preview rendering of the post image has been changed from
inline to a simple URL to the attachment (by default stored in the file system
in a resized and compressed form). This removes the encoding process and allows
the browser to cache the file.

Note that the icon for push notification still use the inline image encoding
because icon are small and it would have been hard to modify as the icon is
stored in a binary field (in that case, the attachment is written after the
post creation and is not available during the creation).

Task-2674439

Part-of: odoo/enterprise#30401