[FIX] documents,web_enterprise: mobile list renderer
During research for odoo/enterprise#30046 , we found out that the
Documents list's selection tests on mobile didn't reflected its actual
usage and behavior.

This commit rewrites parts of this test to properly match the actual
behavior of the documents list selection on mobile.

But, in the process, it also highlighted an issue between the documents'
customization and the list selection on touch devices (ie. long-touch
instead of checkbox to enter in selection mode).

This commit also fixes it by forcing to move the mobile list selection
`includes` up in the asset bundle and ensuring it is applied to
`ListRenderer` before it will be extended.

closes odoo/enterprise#30299

Signed-off-by: Samuel Degueldre <sad@odoo.com>