[FIX] sign: select option deleted when item validate
Steps to reproduce:
- Go to Sign
- Click on a template
- Add a Selection field
- Set the options
- Leave the template
- Go back to it
- Select the field again
- Click Validate again

Current behavior:
The options dissapear from the item

Expected behavior:
The options stay on the item

Explanation:
When we validate and hide the item configuration popup we
pull the item_options from the select input data. The
problem is that item_options is only set with the
selectChangeHandler (meaning when we modify the options),
so if we validate without modifying the options item_options
is never set and the options are deleted. To fix the issue we
also get the options in the start of the popup.

opw-2947995

closes odoo/enterprise#30559

X-original-commit: b0ddc08f77ed0032803e1d22966e617936320db3
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>