[FIX] iot: stop using `get_module_filetree`
The use doesn't really make sense in the first place, the controller
is really just using it as a 2-level listdir.

Could probably be a depth-limited `os.walk`, but a 2-level glob seems
fine as well.

Also lift use of `zipfile` (doesn't seem necessary to iterate twice on
every entry), and set it as compressed as all the content is text
files, seems dumb to just STORE as-is.

closes odoo/enterprise#30403

Related: odoo/odoo#98034
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>