[FIX] knowledge: check user group before throwing access error exception
In the `knowledge.article` model, we manually trigger an access error
when the user is not an admin, does not have "write" permission on the
article and tries to: (1) restore the internal permission of an article,
(2) set the internal permission of an article, (3) set the permission of
a member, (4) add/remove a member from an article.

To check if the user is an admin, we only check the sudo privilege of
the user. In practice, an admin does not neccessarly have sudo privilege
when using the app. As a result, an admin can get an access error while
doing one of the actions cited above.

To make sure that the admin can properly moderate the articles, we will
also check the group of the user before raising an access error exception:
If the user belongs to the `base.group_system`, we will not raise an
access error exception.

task-2893614

X-original-commit: a2ae7b3f141bbd5969fbf27a4f069d792824f1f2
Part-of: odoo/enterprise#30430