[IMP] sale*,partner_com: adapt to invoicing wizard changes.
Now that the invoicing wizard uses `sale.order._prepare_invoice` to
generate the base invoice values, the duplicated code adding the same
invoice values in both the SO invoice values and the wizard invoice
values can be removed.

This commit also adapts an existing test to the new code logic of the
invoicing wizard.

closes odoo/enterprise#30653

Related: odoo/odoo#98600
Related: odoo/upgrade#3799
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>