[REF] spreadsheet: OdooViewsDataSource no longer needs a model
The goal of this commit is to decouple `OdooViewDataSource` from
any underlying model and details about how data is actually loaded.
`OdooViewDataSource` is now focused to know what data to load and to load it
(metadata, search params, domain), regardless of how it's loaded.

The goal is to remove the ListModel and merge it with its data source. Having
two layers is useless.

Part-of: odoo/enterprise#30208