[IMP] knowledge: improve wording and design of the editor view
This commit introduces minor changes regarding the wording and the design
of the editor view of Knowledge. The new wording of the confirmation
dialogs will give more indication of the effect of the action the user
issued. The new design of the share panel will simply make the interface
look prettier.

task-2893088

closes odoo/enterprise#29446

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>