[IMP] l10n_pe_edi: Retrieve CDR if invoice already sent
In PR #23171, we provided functionality such that when sending an
invoice, if the invoice already exists in SUNAT, we catch the error and
retrieve the CDR of the existing invoice.

This was useful in situations where Odoo encounters an exception after
the invoice is sent to SUNAT, which leads to an inconsistent state.

However, that PR only implemented this for the IAP service.

This commit extends this functionality to the SUNAT and Digiflow
services.

This commit also includes the suggestions in PR #29409, which were to
include a check for status code 4000, and to check that the date and
partner of the retrieved CDR match those of the invoice.

closes odoo/enterprise#37736

X-original-commit: 8389445545b4e2e6404e8708a2479ec857bc2132
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>