[FIX] account_accountant: Fix bank reco widget with matching rules using payment tolerance
- Create a reco model with a payment tolerance of 2%
- Create an invoice of 1000
- Create a statement line of 998
=> The whole amount is in write-off

Explanation:
The first issue is the matched aml is mounted inside the widget using 'allow_partial'.
So, the mounted line is a partial of 998 / 1000 instead of the full amount.
The second issue is '_action_select_reconcile_model' is removing all the lines linked to a reconcile model instead of removing only the ones having a different reconcile model.
So, the previously mounted line is removed and then, the write-off is generated for the whole amount.

closes odoo/enterprise#30487

Task: 2929017
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>